Added partial background to some modules
There have been complaints that Four Island is hard to read in the night because all of the text is black. So, for the Blog, whoose only text that is
directly in front of the background is cite text, the cite text has been changed to yellow at night. For the rest, the 4th layout-style center
background has been added back, to greater effect with the fade-out at the top.