Fixed Pingback display problem
This was only noticed when a lot of attention was put on the pingback engine when TimTam's pingback (http://tamasys.wordpress.com/2008/12/03/the-closest-book/) to
(http://fourisland.com/blog/of-books-and-memes/) failed. This in itself was not the actual problem, that being that TimTam seems to have used the incorrect URL to link
to said post, but this was also noticed as rendering the pingbacks themselves as absolutely ugly comments.