Redirected blog post urls
It turns out that FeedBurner somehow gets a hold of the original URLs for blog posts instead of the rewritten ones, so users clicking on them go there. I've fixed the
problem by checking for "index.php" in the URL in blog.php, and redirecting if it finds it. This was the problem with the Pingback thing.