Added support for a variable amount of tags
Previously, the blogging engine only allowed for three tags per post, and it also stored each in seperate fields. Now, all tags are stored in one
field and there can be more than three. The only functionality that has been removed because of this is that now, on archive pages, tags are not
shown next to the current month's posts as they used to be, because the Four Island templating system does not yet support sub-blocks of sub-blocks.