Created a "Mantainence Mode"
A mantainence has been created in the config table so that when it is on, users other than the localhost only see a temporarily down message, while
localhost can still access Four Island. This will be useful during times when changes in this repository do things that require the database schema
to be changed or scripts to be run, such as the previous three changesets.