Fixed BBCode parsing glitch
As previously seen in the 24 Ways post, there was a strange BBCode parsing glitch caused by a few circumstances. There had to be a BBCode tag that
had a parametered definition and a non-parametered definition. Both had to appear in the same post and the parametered had to appear before the
non-parametered.

Because the Four Island BBCode parsing system parses non-parametered tags first, it finds the opening tag of the non-parametered tag first, but when
it tries to find the closing tag, because it is searching from the beginning of the tag, it finds the closing tag of the parametered tag. This causes
strange things to happen.

This has been fixed by forcing the parsing system to start searching for the closing tag after the position where the starting tag is.