Fixed Draft->Pending deletion bug
Previously, if a draft was being moved into the pending queue, it would be deleted instead. The problem turned out to be the fact that when the tag
system was changed (409d99f03471), the single instance in the admin panel of a draft being changed into a pending post was not fixed as the rest
were. The extra column in the "INSERT" query annoyed MySQL as that column no longer exists.