Re-did 404 page
For some reason, the 404 page was previously filled with terrible semantics, IE classes and it didn't look great. Now, the HTML has been improved.
The Search clause has been fixed, there was a bug with displaying the search terms. Also, the error document has been set to the 404 page so the 404
page is actually shown when a page cannot be found.