Eased usability in IE
Previously, IE had a problem with IE 6 that made it impossible to use Four Island. The rightbar was cleared to the bottom, the posts' text were
invisible, the top-fade obscured text and The Fourm was illegiable. This has (mostly) been fixed with a conditional stylesheet. Four Island doesn't
look perfect in IE yet (and probably never will) but at least now it's usable.