Fixed voting issue
Because XHTML was accidentally disabled in the previous changeset, a new error was found that, when in HTML mode (required for IE and layouts prior to 7),
voting did not work due to the encoded ampherstand in the query string. This is odd because what works in XHTML should also work in HTML. So, to fix this,
XHTML was re-enabled and now the ampherstand is only encoded when XHTML mode is on.

Also fixed some positioning errors with the tails of bubbles.