Removed <[!CDATA]>s from admin panel
When Fouripedia compatibility was added, the admin panel was changed from HTML to XHTML, and so all CDATAs appeared in textboxes instead of doing their
job. They have been removed, but this is only a temporary fix as the proper solution would be to re-enable XHTML on the admin panel.