Organized blog module
Previously, welcome.php was always loaded if no module was specified. Because of this, the index page blog section was there, while the rest of the blog
code (plus archive code) was in blog.php. Now, the all of the blog is in blog.php and all of the archive is in archive.php.