Removed XHTML
XHTML was causing so many problems that it was simply unbearable. Four Island 3.0 will be sure to have no traces left of this annoyance.

Also, a lot of files are showing up as being changed because this commit was on the server and I had to modifiy permissions for files (all of the files that were new since my
last server commit) so I could commit.