Fixed (temporarily) Memorial Day's date
Because Memorial Day falls on the LAST Monday of May, rather than a specific week number like the fourth week, and May can sometimes have four mondays
and sometimes have five mondays, this is quite dangerous because in 2012, May will only have four Mondays, which will render this code invalid. However,
Four Island 3.0 should hopefully be out by 2012, which will hopefully not contain this bug (a.k.a. by allowing a date to be specified as a the last day of
a month as well as during a specific week).